+++ src/frontend/app/app.router.ts
@@ -34,7 +34,9 @@ export const appRoutes: Routes = [
 
 @NgModule({
     imports: [
-        RouterModule.forRoot(appRoutes)
+        RouterModule.forRoot(appRoutes,
+            // { enableTracing: true } // <-- debugging purposes only
+        ),
     ],
     exports: [
         RouterModule