+++ src/frontend/app/services/custom-url-serializer.ts
@@ -1,0 +1,17 @@
+import { DefaultUrlSerializer, UrlSerializer, UrlTree } from '@angular/router';
+
+// seen on https://stackoverflow.com/questions/39541185/custom-encoding-for-urls-using-angular-2-router-using-a-sign-in-place-of-a-sp
+export default class CustomUrlSerializer implements UrlSerializer {
+    private _defaultUrlSerializer: DefaultUrlSerializer = new DefaultUrlSerializer();
+
+    parse(url: string): UrlTree {
+        // Encode parentheses
+        url = url.replace(/\(/g, '%28').replace(/\)/g, '%29');
+        // Use the default serializer.
+        return this._defaultUrlSerializer.parse(url)
+    }
+
+    serialize(tree: UrlTree): string {
+        return this._defaultUrlSerializer.serialize(tree).replace(/%28/g, '(').replace(/%29/g, ')');
+    }
+}